+++ src/engine/components/numericLineGenerator/index.stories.jsx
@@ -9,7 +9,7 @@ import { storiesOf } from '@kadira/storybook';
 
 import PixiWrapper from 'test/PixiWrapper';
 import numberGenerator from 'engine/components/numberGenerator';
-import { TEN, HALF, ONE } from 'constants/numbers';
+import { TEN, HALF, ZERO } from 'constants/numbers';
 
 import arrowGenerator from './arrowGenerator';
 import lineGenerator from './lineGenerator';
@@ -24,8 +24,12 @@ const WIDTH_LINE = 640;
 const HEIGHT_LINE = 200;
 const QUARTER = 4;
 const HEIGHT = 8;
+const NINE = 9;
 
 storiesOf('engine.components.numericLine', module)
+  //////////////////////////////////
+  // Arrows
+  //////////////////////////////////
   .add('simple left arrow', () => {
     const arrow = arrowGenerator();
 
@@ -42,6 +46,9 @@ storiesOf('engine.components.numericLine', module)
 
     return <PixiWrapper component={arrow} height={HEIGHT_ARROW} width={WIDTH_ARROW} />;
   })
+  //////////////////////////////////
+  // Line
+  //////////////////////////////////
   .add('simple line', () => {
     const line = lineGenerator(HEIGHT);
 
@@ -50,26 +57,57 @@ storiesOf('engine.components.numericLine', module)
 
     return <PixiWrapper component={line.view} height={HEIGHT_LINE} width={WIDTH_LINE} />;
   })
-  .add('line with a number', () => {
+  .add('line with static numbers', () => {
+    const line = lineGenerator(HEIGHT, [
+      [0, 0], [1, 1], [2, 2], [3, 3], [4, 4], [5, 5], [6, 6], [7, 7], // eslint-disable-line no-magic-numbers
+    ]);
+
+    line.view.x = 32;
+    line.view.y = HEIGHT_LINE / QUARTER;
+
+    return (
+      <PixiWrapper component={line.view} height={HEIGHT_LINE} width={WIDTH_LINE} />
+    );
+  })
+  .add('line with an animated number', () => {
     const line = lineGenerator(HEIGHT);
     const ten = numberGenerator(TEN, '0,0');
 
     line.view.x = 32;
     line.view.y = HEIGHT_LINE / QUARTER;
 
-    line.receiveNumberAtPosition(ten, ONE);
+    line.receiveNumberAtPosition(ten, ZERO);
 
     return (
       <PixiWrapper component={line.view} height={HEIGHT_LINE} width={WIDTH_LINE} />
     );
   })
+  //////////////////////////////////
+  // Numeric Line
+  //////////////////////////////////
   .add('basic numeric line', () => {
-    const numericLine1 = numericLineGenerator(TEN);
+    const numericLine = numericLineGenerator(TEN);
 
-    numericLine1.view.x = (WIDTH_NUMERIC_LINE - numericLine1.view.width) / HALF;
-    numericLine1.view.y = (HEIGHT_NUMERIC_LINE - numericLine1.view.height) / HALF;
+    numericLine.view.x = (WIDTH_NUMERIC_LINE - numericLine.view.width) / HALF;
+    numericLine.view.y = (HEIGHT_NUMERIC_LINE - numericLine.view.height) / HALF;
 
     return (
-      <PixiWrapper component={numericLine1.view} height={HEIGHT_NUMERIC_LINE} width={WIDTH_NUMERIC_LINE} />
+      <PixiWrapper component={numericLine.view} height={HEIGHT_NUMERIC_LINE} width={WIDTH_NUMERIC_LINE} />
+    );
+  })
+  .add('numeric line with animated and static numbers', () => {
+    const numericLine = numericLineGenerator(TEN, [
+      // eslint-disable-next-line no-magic-numbers
+      [0, 1], [1, 2], [2, 3], [3, 4], [4, 5], [5, 6], [6, 7], [7, 8], [8, 9],
+    ]);
+    const ten = numberGenerator(TEN, '0,0');
+
+    numericLine.receiveNumberAtPosition(ten, NINE);
+
+    numericLine.view.x = (WIDTH_NUMERIC_LINE - numericLine.view.width) / HALF;
+    numericLine.view.y = (HEIGHT_NUMERIC_LINE - numericLine.view.height) / HALF;
+
+    return (
+      <PixiWrapper component={numericLine.view} height={HEIGHT_NUMERIC_LINE} width={WIDTH_NUMERIC_LINE} />
     );
   });